Split body concent counter from grader
I have the body concentration counting part as an individual object,
and move it out with the FaceExistenceRateCounter to another file.
