+++ intervals.json
@@ -1,42 +1,217 @@
 [
   {
-    "start": 1639972441,
-    "end": 1639972501,
-    "grade": 0.84
+    "start": 1640399701,
+    "end": 1640399761,
+    "grade": 0.89
   },
   {
-    "start": 1639972509,
-    "end": 1639972569,
-    "grade": 0.88
+    "start": 1640399761,
+    "end": 1640399821,
+    "grade": 0.81
   },
   {
-    "start": 1639972592,
-    "end": 1639972652,
-    "grade": 0.62
-  },
-  {
-    "start": 1639972652,
-    "end": 1639972712,
-    "grade": 0.0
-  },
-  {
-    "start": 1639972712,
-    "end": 1639972772,
-    "grade": 0.71
-  },
-  {
-    "start": 1639972830,
-    "end": 1639972890,
+    "start": 1640399821,
+    "end": 1640399881,
     "grade": 1.0
   },
   {
-    "start": 1639972890,
-    "end": 1639972950,
-    "grade": 0.8
+    "start": 1640399881,
+    "end": 1640399941,
+    "grade": 1.0
   },
   {
-    "start": 1639972950,
-    "end": 1639973010,
-    "grade": 0.17
+    "start": 1640399941,
+    "end": 1640399981,
+    "grade": 0.77
+  },
+  {
+    "start": 1640399981,
+    "end": 1640400013,
+    "grade": 0.03
+  },
+  {
+    "start": 1640400035,
+    "end": 1640400095,
+    "grade": 0.65
+  },
+  {
+    "start": 1640400095,
+    "end": 1640400155,
+    "grade": 0.66
+  },
+  {
+    "start": 1640400155,
+    "end": 1640400215,
+    "grade": 0.88
+  },
+  {
+    "start": 1640400215,
+    "end": 1640400275,
+    "grade": 0.92
+  },
+  {
+    "start": 1640400275,
+    "end": 1640400335,
+    "grade": 0.0
+  },
+  {
+    "start": 1640400335,
+    "end": 1640400395,
+    "grade": 0.0
+  },
+  {
+    "start": 1640400395,
+    "end": 1640400455,
+    "grade": 0.36
+  },
+  {
+    "start": 1640400455,
+    "end": 1640400515,
+    "grade": 0.85
+  },
+  {
+    "start": 1640400515,
+    "end": 1640400575,
+    "grade": 0.95
+  },
+  {
+    "start": 1640400575,
+    "end": 1640400635,
+    "grade": 0.98
+  },
+  {
+    "start": 1640400635,
+    "end": 1640400695,
+    "grade": 1.0
+  },
+  {
+    "start": 1640400695,
+    "end": 1640400755,
+    "grade": 0.64
+  },
+  {
+    "start": 1640400755,
+    "end": 1640400815,
+    "grade": 0.03
+  },
+  {
+    "start": 1640400815,
+    "end": 1640400875,
+    "grade": 0.1
+  },
+  {
+    "start": 1640400875,
+    "end": 1640400935,
+    "grade": 0.01
+  },
+  {
+    "start": 1640400944,
+    "end": 1640401004,
+    "grade": 0.61
+  },
+  {
+    "start": 1640401004,
+    "end": 1640401064,
+    "grade": 0.2
+  },
+  {
+    "start": 1640401064,
+    "end": 1640401124,
+    "grade": 0.54
+  },
+  {
+    "start": 1640401124,
+    "end": 1640401184,
+    "grade": 1.0
+  },
+  {
+    "start": 1640401184,
+    "end": 1640401244,
+    "grade": 0.96
+  },
+  {
+    "start": 1640401255,
+    "end": 1640401315,
+    "grade": 0.72
+  },
+  {
+    "start": 1640401315,
+    "end": 1640401345,
+    "grade": 0.66
+  },
+  {
+    "start": 1640401345,
+    "end": 1640401405,
+    "grade": 0.81
+  },
+  {
+    "start": 1640401408,
+    "end": 1640401468,
+    "grade": 0.61
+  },
+  {
+    "start": 1640401480,
+    "end": 1640401540,
+    "grade": 0.4
+  },
+  {
+    "start": 1640401540,
+    "end": 1640401600,
+    "grade": 0.54
+  },
+  {
+    "start": 1640401600,
+    "end": 1640401660,
+    "grade": 0.38
+  },
+  {
+    "start": 1640401660,
+    "end": 1640401720,
+    "grade": 0.49
+  },
+  {
+    "start": 1640401741,
+    "end": 1640401801,
+    "grade": 0.65
+  },
+  {
+    "start": 1640401805,
+    "end": 1640401865,
+    "grade": 0.7
+  },
+  {
+    "start": 1640401869,
+    "end": 1640401929,
+    "grade": 0.61
+  },
+  {
+    "start": 1640401929,
+    "end": 1640401989,
+    "grade": 0.44
+  },
+  {
+    "start": 1640401989,
+    "end": 1640402019,
+    "grade": 0.4
+  },
+  {
+    "start": 1640402026,
+    "end": 1640402086,
+    "grade": 0.6
+  },
+  {
+    "start": 1640402115,
+    "end": 1640402175,
+    "grade": 0.61
+  },
+  {
+    "start": 1640402175,
+    "end": 1640402235,
+    "grade": 0.63
+  },
+  {
+    "start": 1640402245,
+    "end": 1640402305,
+    "grade": 0.63
   }
 ]
